@@ -18,12 +18,12 @@
 		}
 	],
 	"require": {
-		"php": ">=7.4",
-		"cakephp/cakephp": "^4.2.0"
+		"php": ">=8.1",
+		"cakephp/cakephp": "5.x-dev"
 	},
 	"require-dev": {
-		"phpunit/phpunit": "^9.5",
-		"fig-r/psr2r-sniffer": "dev-master",
+		"phpunit/phpunit": "^9.5.16",
+		"fig-r/psr2r-sniffer": "dev-next",
 		"league/commonmark": "^1.5 || ^2.0",
 		"mjohnson/decoda": "^6.12",
 		"dereuromark/media-embed": "^0.6"
@@ -56,5 +56,6 @@
 		"allow-plugins": {
 			"dealerdirect/phpcodesniffer-composer-installer": true
 		}
-	}
+	},
+	"minimum-stability": "dev"
 }
