@@ -23,7 +23,7 @@ class CommonMarkMarkdown implements MarkdownInterface {
 	/**
 	 * @var array<string, mixed>
 	 */
-	protected $_defaultConfig = [
+	protected array $_defaultConfig = [
 	];
 
 	/**
